Anpassungen für Los E
git-svn-id: svn://ntoserver/Datenverteiler/VRZ3/trunk/de.bsvrz.sys.funclib.bitctrl@18263 40c0be50-5d29-0410-901d-fb59edb5cafc

+++ junit/de/bsvrz/sys/funclib/bitctrl/util/TestSortierteListe.java
@@ -168,7 +168,7 @@ public class TestSortierteListe {
 	 */
 	@Test
 	public void testFirst() {
-		assertEquals(0, liste.first());
+		assertEquals((Integer)0, liste.first());
 	}
 
 	/**
@@ -192,7 +192,7 @@ public class TestSortierteListe {
 	 */
 	@Test
 	public void testLast() {
-		assertEquals(9, liste.last());
+		assertEquals((Integer)9, liste.last());
 	}
 
 	/**
